added form based authentication

+++ csharp/DemoProject/Controllers/EmployeeController.cs
@@ -4,12 +4,74 @@ using System.Data.Entity;
 using System.Linq;
 using System.Web;
 using System.Web.Mvc;
+using System.Web.Security;
 using DemoProject.Models;
 
 namespace DemoProject.Controllers
 {
+    [Authorize]
     public class EmployeeController : Controller
     {
+        [AllowAnonymous]
+        [Route("login")]
+        public ActionResult Login()
+        {
+            return View();
+        }
+
+        [AllowAnonymous]
+        [HttpPost]
+        [Route("login")]
+        public ActionResult Login([Bind(Include ="Email,Password")] User user)
+        {
+            using (var context=new UserContext())
+            {
+                bool IsValidEmployee = context.Users.Any(u => u.Email.Equals(user.Email) && u.Password.Equals(user.Password));
+                if (IsValidEmployee)
+                {
+              
+                    FormsAuthentication.SetAuthCookie(user.Email, false);
+                    return RedirectToAction("GetAllEmployees");
+ 
+
+                }
+                ModelState.AddModelError("", "invalid Username or Password");
+              
+            }
+            return View();
+        }
+        [AllowAnonymous]
+        [Route("signup")]
+        public ActionResult SignUp()
+        {
+            return View();
+        }
+
+        [AllowAnonymous]
+        [HttpPost]
+        [Route("signup")]
+        public ActionResult SignUp(User user)
+        {
+            if (ModelState.IsValid)
+            {
+                using (var context = new UserContext()) {
+              
+                    context.Users.Add(user);
+                    context.SaveChanges();
+                }
+              
+            }
+
+            return RedirectToAction("Login");
+        }
+
+        [Route("logout")]
+        public ActionResult Logout()
+        {
+            FormsAuthentication.SignOut();
+            return RedirectToAction("Login");
+        }
+
         // Create Employee - GET
         [HttpGet]
         [Route("create")]
@@ -20,17 +82,20 @@ namespace DemoProject.Controllers
 
         // Create Employee - POST
         [HttpPost]
-        [Route("create/{employee}")]
+        [Route("create")]
         public ActionResult CreateEmployee(Employee employee)
         {
-            if (ModelState.IsValid)
+            if (User.Identity.IsAuthenticated)
             {
-                using (var context = new EmployeeContext())
+                if (ModelState.IsValid)
                 {
-                    context.Employees.Add(employee);
-                    context.SaveChanges();
-                }
+                    using (var context = new EmployeeContext())
+                    {
+                        context.Employees.Add(employee);
+                        context.SaveChanges();
+                    }
 
+                }
             }
 
             return RedirectToAction("GetAllEmployees");
@@ -39,31 +104,45 @@ namespace DemoProject.Controllers
         // Get All Employees
         [HttpGet]
         [Route("employees")]
+        [AllowAnonymous]
         public ActionResult GetAllEmployees()
         {
-            List<Employee> employees;
-            using (var context = new EmployeeContext())
+            if (User.Identity.IsAuthenticated)
             {
-                employees = context.Employees.ToList();
+                List<Employee> employees;
+                using (var context = new EmployeeContext())
+                {
+                    employees = context.Employees.ToList();
+                }
+                ViewBag.Employees = employees;
+
             }
-            ViewBag.Employees = employees;
-            return View();
+            else
+            {
+                return RedirectToAction("Login");
+            }
+                return View();
         }
+
 
         // Delete Employee
         [HttpPost]
         [Route("delete/{id}")]
         public ActionResult Delete(int id)
         {
-            using (var context = new EmployeeContext())
+            if (User.Identity.IsAuthenticated)
             {
-                var existingEmployee = context.Employees.Find(id);
-                if (existingEmployee != null)
+                using (var context = new EmployeeContext())
                 {
-                    context.Employees.Remove(existingEmployee);
-                    context.SaveChanges();
+                    var existingEmployee = context.Employees.Find(id);
+                    if (existingEmployee != null)
+                    {
+                        context.Employees.Remove(existingEmployee);
+                        context.SaveChanges();
+                    }
                 }
             }
+           
 
             return RedirectToAction("GetAllEmployees");
         }
